@@ -1,12 +1,12 @@
 <?php
 // TCPDF FONT FILE DESCRIPTION
-$type='core';
-$name='Courier';
-$up=-100;
-$ut=50;
-$dw=600;
-$diff='';
-$enc='';
-$desc=array('Flags'=>33,'FontBBox'=>'[-23 -250 715 805]','ItalicAngle'=>0,'Ascent'=>805,'Descent'=>-250,'Leading'=>0,'CapHeight'=>562,'XHeight'=>426,'StemV'=>51,'StemH'=>51,'AvgWidth'=>600,'MaxWidth'=>600,'MissingWidth'=>600);
-$cw=array(0=>600,1=>600,2=>600,3=>600,4=>600,5=>600,6=>600,7=>600,8=>600,9=>600,10=>600,11=>600,12=>600,13=>600,14=>600,15=>600,16=>600,17=>600,18=>600,19=>600,20=>600,21=>600,22=>600,23=>600,24=>600,25=>600,26=>600,27=>600,28=>600,29=>600,30=>600,31=>600,32=>600,33=>600,34=>600,35=>600,36=>600,37=>600,38=>600,39=>600,40=>600,41=>600,42=>600,43=>600,44=>600,45=>600,46=>600,47=>600,48=>600,49=>600,50=>600,51=>600,52=>600,53=>600,54=>600,55=>600,56=>600,57=>600,58=>600,59=>600,60=>600,61=>600,62=>600,63=>600,64=>600,65=>600,66=>600,67=>600,68=>600,69=>600,70=>600,71=>600,72=>600,73=>600,74=>600,75=>600,76=>600,77=>600,78=>600,79=>600,80=>600,81=>600,82=>600,83=>600,84=>600,85=>600,86=>600,87=>600,88=>600,89=>600,90=>600,91=>600,92=>600,93=>600,94=>600,95=>600,96=>600,97=>600,98=>600,99=>600,100=>600,101=>600,102=>600,103=>600,104=>600,105=>600,106=>600,107=>600,108=>600,109=>600,110=>600,111=>600,112=>600,113=>600,114=>600,115=>600,116=>600,117=>600,118=>600,119=>600,120=>600,121=>600,122=>600,123=>600,124=>600,125=>600,126=>600,127=>600,128=>600,129=>600,130=>600,131=>600,132=>600,133=>600,134=>600,135=>600,136=>600,137=>600,138=>600,139=>600,140=>600,141=>600,142=>600,143=>600,144=>600,145=>600,146=>600,147=>600,148=>600,149=>600,150=>600,151=>600,152=>600,153=>600,154=>600,155=>600,156=>600,157=>600,158=>600,159=>600,160=>600,161=>600,162=>600,163=>600,164=>600,165=>600,166=>600,167=>600,168=>600,169=>600,170=>600,171=>600,172=>600,173=>600,174=>600,175=>600,176=>600,177=>600,178=>600,179=>600,180=>600,181=>600,182=>600,183=>600,184=>600,185=>600,186=>600,187=>600,188=>600,189=>600,190=>600,191=>600,192=>600,193=>600,194=>600,195=>600,196=>600,197=>600,198=>600,199=>600,200=>600,201=>600,202=>600,203=>600,204=>600,205=>600,206=>600,207=>600,208=>600,209=>600,210=>600,211=>600,212=>600,213=>600,214=>600,215=>600,216=>600,217=>600,218=>600,219=>600,220=>600,221=>600,222=>600,223=>600,224=>600,225=>600,226=>600,227=>600,228=>600,229=>600,230=>600,231=>600,232=>600,233=>600,234=>600,235=>600,236=>600,237=>600,238=>600,239=>600,240=>600,241=>600,242=>600,243=>600,244=>600,245=>600,246=>600,247=>600,248=>600,249=>600,250=>600,251=>600,252=>600,253=>600,254=>600,255=>600);
+$type = 'core';
+$name = 'Courier';
+$up = -100;
+$ut = 50;
+$dw = 600;
+$diff = '';
+$enc = '';
+$desc = ['Flags' => 33,'FontBBox' => '[-23 -250 715 805]','ItalicAngle' => 0,'Ascent' => 805,'Descent' => -250,'Leading' => 0,'CapHeight' => 562,'XHeight' => 426,'StemV' => 51,'StemH' => 51,'AvgWidth' => 600,'MaxWidth' => 600,'MissingWidth' => 600];
+$cw = [0 => 600,1 => 600,2 => 600,3 => 600,4 => 600,5 => 600,6 => 600,7 => 600,8 => 600,9 => 600,10 => 600,11 => 600,12 => 600,13 => 600,14 => 600,15 => 600,16 => 600,17 => 600,18 => 600,19 => 600,20 => 600,21 => 600,22 => 600,23 => 600,24 => 600,25 => 600,26 => 600,27 => 600,28 => 600,29 => 600,30 => 600,31 => 600,32 => 600,33 => 600,34 => 600,35 => 600,36 => 600,37 => 600,38 => 600,39 => 600,40 => 600,41 => 600,42 => 600,43 => 600,44 => 600,45 => 600,46 => 600,47 => 600,48 => 600,49 => 600,50 => 600,51 => 600,52 => 600,53 => 600,54 => 600,55 => 600,56 => 600,57 => 600,58 => 600,59 => 600,60 => 600,61 => 600,62 => 600,63 => 600,64 => 600,65 => 600,66 => 600,67 => 600,68 => 600,69 => 600,70 => 600,71 => 600,72 => 600,73 => 600,74 => 600,75 => 600,76 => 600,77 => 600,78 => 600,79 => 600,80 => 600,81 => 600,82 => 600,83 => 600,84 => 600,85 => 600,86 => 600,87 => 600,88 => 600,89 => 600,90 => 600,91 => 600,92 => 600,93 => 600,94 => 600,95 => 600,96 => 600,97 => 600,98 => 600,99 => 600,100 => 600,101 => 600,102 => 600,103 => 600,104 => 600,105 => 600,106 => 600,107 => 600,108 => 600,109 => 600,110 => 600,111 => 600,112 => 600,113 => 600,114 => 600,115 => 600,116 => 600,117 => 600,118 => 600,119 => 600,120 => 600,121 => 600,122 => 600,123 => 600,124 => 600,125 => 600,126 => 600,127 => 600,128 => 600,129 => 600,130 => 600,131 => 600,132 => 600,133 => 600,134 => 600,135 => 600,136 => 600,137 => 600,138 => 600,139 => 600,140 => 600,141 => 600,142 => 600,143 => 600,144 => 600,145 => 600,146 => 600,147 => 600,148 => 600,149 => 600,150 => 600,151 => 600,152 => 600,153 => 600,154 => 600,155 => 600,156 => 600,157 => 600,158 => 600,159 => 600,160 => 600,161 => 600,162 => 600,163 => 600,164 => 600,165 => 600,166 => 600,167 => 600,168 => 600,169 => 600,170 => 600,171 => 600,172 => 600,173 => 600,174 => 600,175 => 600,176 => 600,177 => 600,178 => 600,179 => 600,180 => 600,181 => 600,182 => 600,183 => 600,184 => 600,185 => 600,186 => 600,187 => 600,188 => 600,189 => 600,190 => 600,191 => 600,192 => 600,193 => 600,194 => 600,195 => 600,196 => 600,197 => 600,198 => 600,199 => 600,200 => 600,201 => 600,202 => 600,203 => 600,204 => 600,205 => 600,206 => 600,207 => 600,208 => 600,209 => 600,210 => 600,211 => 600,212 => 600,213 => 600,214 => 600,215 => 600,216 => 600,217 => 600,218 => 600,219 => 600,220 => 600,221 => 600,222 => 600,223 => 600,224 => 600,225 => 600,226 => 600,227 => 600,228 => 600,229 => 600,230 => 600,231 => 600,232 => 600,233 => 600,234 => 600,235 => 600,236 => 600,237 => 600,238 => 600,239 => 600,240 => 600,241 => 600,242 => 600,243 => 600,244 => 600,245 => 600,246 => 600,247 => 600,248 => 600,249 => 600,250 => 600,251 => 600,252 => 600,253 => 600,254 => 600,255 => 600];
 // --- EOF ---
