@@ -1,12 +1,12 @@
 <?php
 // TCPDF FONT FILE DESCRIPTION
-$type='core';
-$name='Helvetica-Bold';
-$up=-100;
-$ut=50;
-$dw=535;
-$diff='';
-$enc='';
-$desc=array('Flags'=>32,'FontBBox'=>'[-170 -228 1003 962]','ItalicAngle'=>0,'Ascent'=>962,'Descent'=>-228,'Leading'=>0,'CapHeight'=>718,'XHeight'=>532,'StemV'=>140,'StemH'=>118,'AvgWidth'=>535,'MaxWidth'=>1000,'MissingWidth'=>535);
-$cw=array(0=>278,1=>278,2=>278,3=>278,4=>278,5=>278,6=>278,7=>278,8=>278,9=>278,10=>278,11=>278,12=>278,13=>278,14=>278,15=>278,16=>278,17=>278,18=>278,19=>278,20=>278,21=>278,22=>278,23=>278,24=>278,25=>278,26=>278,27=>278,28=>278,29=>278,30=>278,31=>278,32=>278,33=>333,34=>474,35=>556,36=>556,37=>889,38=>722,39=>238,40=>333,41=>333,42=>389,43=>584,44=>278,45=>333,46=>278,47=>278,48=>556,49=>556,50=>556,51=>556,52=>556,53=>556,54=>556,55=>556,56=>556,57=>556,58=>333,59=>333,60=>584,61=>584,62=>584,63=>611,64=>975,65=>722,66=>722,67=>722,68=>722,69=>667,70=>611,71=>778,72=>722,73=>278,74=>556,75=>722,76=>611,77=>833,78=>722,79=>778,80=>667,81=>778,82=>722,83=>667,84=>611,85=>722,86=>667,87=>944,88=>667,89=>667,90=>611,91=>333,92=>278,93=>333,94=>584,95=>556,96=>333,97=>556,98=>611,99=>556,100=>611,101=>556,102=>333,103=>611,104=>611,105=>278,106=>278,107=>556,108=>278,109=>889,110=>611,111=>611,112=>611,113=>611,114=>389,115=>556,116=>333,117=>611,118=>556,119=>778,120=>556,121=>556,122=>500,123=>389,124=>280,125=>389,126=>584,127=>350,128=>556,129=>350,130=>278,131=>556,132=>500,133=>1000,134=>556,135=>556,136=>333,137=>1000,138=>667,139=>333,140=>1000,141=>350,142=>611,143=>350,144=>350,145=>278,146=>278,147=>500,148=>500,149=>350,150=>556,151=>1000,152=>333,153=>1000,154=>556,155=>333,156=>944,157=>350,158=>500,159=>667,160=>278,161=>333,162=>556,163=>556,164=>556,165=>556,166=>280,167=>556,168=>333,169=>737,170=>370,171=>556,172=>584,173=>333,174=>737,175=>333,176=>400,177=>584,178=>333,179=>333,180=>333,181=>611,182=>556,183=>278,184=>333,185=>333,186=>365,187=>556,188=>834,189=>834,190=>834,191=>611,192=>722,193=>722,194=>722,195=>722,196=>722,197=>722,198=>1000,199=>722,200=>667,201=>667,202=>667,203=>667,204=>278,205=>278,206=>278,207=>278,208=>722,209=>722,210=>778,211=>778,212=>778,213=>778,214=>778,215=>584,216=>778,217=>722,218=>722,219=>722,220=>722,221=>667,222=>667,223=>611,224=>556,225=>556,226=>556,227=>556,228=>556,229=>556,230=>889,231=>556,232=>556,233=>556,234=>556,235=>556,236=>278,237=>278,238=>278,239=>278,240=>611,241=>611,242=>611,243=>611,244=>611,245=>611,246=>611,247=>584,248=>611,249=>611,250=>611,251=>611,252=>611,253=>556,254=>611,255=>556);
+$type = 'core';
+$name = 'Helvetica-Bold';
+$up = -100;
+$ut = 50;
+$dw = 535;
+$diff = '';
+$enc = '';
+$desc = ['Flags' => 32,'FontBBox' => '[-170 -228 1003 962]','ItalicAngle' => 0,'Ascent' => 962,'Descent' => -228,'Leading' => 0,'CapHeight' => 718,'XHeight' => 532,'StemV' => 140,'StemH' => 118,'AvgWidth' => 535,'MaxWidth' => 1000,'MissingWidth' => 535];
+$cw = [0 => 278,1 => 278,2 => 278,3 => 278,4 => 278,5 => 278,6 => 278,7 => 278,8 => 278,9 => 278,10 => 278,11 => 278,12 => 278,13 => 278,14 => 278,15 => 278,16 => 278,17 => 278,18 => 278,19 => 278,20 => 278,21 => 278,22 => 278,23 => 278,24 => 278,25 => 278,26 => 278,27 => 278,28 => 278,29 => 278,30 => 278,31 => 278,32 => 278,33 => 333,34 => 474,35 => 556,36 => 556,37 => 889,38 => 722,39 => 238,40 => 333,41 => 333,42 => 389,43 => 584,44 => 278,45 => 333,46 => 278,47 => 278,48 => 556,49 => 556,50 => 556,51 => 556,52 => 556,53 => 556,54 => 556,55 => 556,56 => 556,57 => 556,58 => 333,59 => 333,60 => 584,61 => 584,62 => 584,63 => 611,64 => 975,65 => 722,66 => 722,67 => 722,68 => 722,69 => 667,70 => 611,71 => 778,72 => 722,73 => 278,74 => 556,75 => 722,76 => 611,77 => 833,78 => 722,79 => 778,80 => 667,81 => 778,82 => 722,83 => 667,84 => 611,85 => 722,86 => 667,87 => 944,88 => 667,89 => 667,90 => 611,91 => 333,92 => 278,93 => 333,94 => 584,95 => 556,96 => 333,97 => 556,98 => 611,99 => 556,100 => 611,101 => 556,102 => 333,103 => 611,104 => 611,105 => 278,106 => 278,107 => 556,108 => 278,109 => 889,110 => 611,111 => 611,112 => 611,113 => 611,114 => 389,115 => 556,116 => 333,117 => 611,118 => 556,119 => 778,120 => 556,121 => 556,122 => 500,123 => 389,124 => 280,125 => 389,126 => 584,127 => 350,128 => 556,129 => 350,130 => 278,131 => 556,132 => 500,133 => 1000,134 => 556,135 => 556,136 => 333,137 => 1000,138 => 667,139 => 333,140 => 1000,141 => 350,142 => 611,143 => 350,144 => 350,145 => 278,146 => 278,147 => 500,148 => 500,149 => 350,150 => 556,151 => 1000,152 => 333,153 => 1000,154 => 556,155 => 333,156 => 944,157 => 350,158 => 500,159 => 667,160 => 278,161 => 333,162 => 556,163 => 556,164 => 556,165 => 556,166 => 280,167 => 556,168 => 333,169 => 737,170 => 370,171 => 556,172 => 584,173 => 333,174 => 737,175 => 333,176 => 400,177 => 584,178 => 333,179 => 333,180 => 333,181 => 611,182 => 556,183 => 278,184 => 333,185 => 333,186 => 365,187 => 556,188 => 834,189 => 834,190 => 834,191 => 611,192 => 722,193 => 722,194 => 722,195 => 722,196 => 722,197 => 722,198 => 1000,199 => 722,200 => 667,201 => 667,202 => 667,203 => 667,204 => 278,205 => 278,206 => 278,207 => 278,208 => 722,209 => 722,210 => 778,211 => 778,212 => 778,213 => 778,214 => 778,215 => 584,216 => 778,217 => 722,218 => 722,219 => 722,220 => 722,221 => 667,222 => 667,223 => 611,224 => 556,225 => 556,226 => 556,227 => 556,228 => 556,229 => 556,230 => 889,231 => 556,232 => 556,233 => 556,234 => 556,235 => 556,236 => 278,237 => 278,238 => 278,239 => 278,240 => 611,241 => 611,242 => 611,243 => 611,244 => 611,245 => 611,246 => 611,247 => 584,248 => 611,249 => 611,250 => 611,251 => 611,252 => 611,253 => 556,254 => 611,255 => 556];
 // --- EOF ---
