@@ -1,12 +1,12 @@
 <?php
 // TCPDF FONT FILE DESCRIPTION
-$type='core';
-$name='Helvetica-Oblique';
-$up=-100;
-$ut=50;
-$dw=513;
-$diff='';
-$enc='';
-$desc=array('Flags'=>96,'FontBBox'=>'[-170 -225 1116 931]','ItalicAngle'=>-12,'Ascent'=>931,'Descent'=>-225,'Leading'=>0,'CapHeight'=>718,'XHeight'=>523,'StemV'=>88,'StemH'=>76,'AvgWidth'=>513,'MaxWidth'=>1015,'MissingWidth'=>513);
-$cw=array(0=>278,1=>278,2=>278,3=>278,4=>278,5=>278,6=>278,7=>278,8=>278,9=>278,10=>278,11=>278,12=>278,13=>278,14=>278,15=>278,16=>278,17=>278,18=>278,19=>278,20=>278,21=>278,22=>278,23=>278,24=>278,25=>278,26=>278,27=>278,28=>278,29=>278,30=>278,31=>278,32=>278,33=>278,34=>355,35=>556,36=>556,37=>889,38=>667,39=>191,40=>333,41=>333,42=>389,43=>584,44=>278,45=>333,46=>278,47=>278,48=>556,49=>556,50=>556,51=>556,52=>556,53=>556,54=>556,55=>556,56=>556,57=>556,58=>278,59=>278,60=>584,61=>584,62=>584,63=>556,64=>1015,65=>667,66=>667,67=>722,68=>722,69=>667,70=>611,71=>778,72=>722,73=>278,74=>500,75=>667,76=>556,77=>833,78=>722,79=>778,80=>667,81=>778,82=>722,83=>667,84=>611,85=>722,86=>667,87=>944,88=>667,89=>667,90=>611,91=>278,92=>278,93=>278,94=>469,95=>556,96=>333,97=>556,98=>556,99=>500,100=>556,101=>556,102=>278,103=>556,104=>556,105=>222,106=>222,107=>500,108=>222,109=>833,110=>556,111=>556,112=>556,113=>556,114=>333,115=>500,116=>278,117=>556,118=>500,119=>722,120=>500,121=>500,122=>500,123=>334,124=>260,125=>334,126=>584,127=>350,128=>556,129=>350,130=>222,131=>556,132=>333,133=>1000,134=>556,135=>556,136=>333,137=>1000,138=>667,139=>333,140=>1000,141=>350,142=>611,143=>350,144=>350,145=>222,146=>222,147=>333,148=>333,149=>350,150=>556,151=>1000,152=>333,153=>1000,154=>500,155=>333,156=>944,157=>350,158=>500,159=>667,160=>278,161=>333,162=>556,163=>556,164=>556,165=>556,166=>260,167=>556,168=>333,169=>737,170=>370,171=>556,172=>584,173=>333,174=>737,175=>333,176=>400,177=>584,178=>333,179=>333,180=>333,181=>556,182=>537,183=>278,184=>333,185=>333,186=>365,187=>556,188=>834,189=>834,190=>834,191=>611,192=>667,193=>667,194=>667,195=>667,196=>667,197=>667,198=>1000,199=>722,200=>667,201=>667,202=>667,203=>667,204=>278,205=>278,206=>278,207=>278,208=>722,209=>722,210=>778,211=>778,212=>778,213=>778,214=>778,215=>584,216=>778,217=>722,218=>722,219=>722,220=>722,221=>667,222=>667,223=>611,224=>556,225=>556,226=>556,227=>556,228=>556,229=>556,230=>889,231=>500,232=>556,233=>556,234=>556,235=>556,236=>278,237=>278,238=>278,239=>278,240=>556,241=>556,242=>556,243=>556,244=>556,245=>556,246=>556,247=>584,248=>611,249=>556,250=>556,251=>556,252=>556,253=>500,254=>556,255=>500);
+$type = 'core';
+$name = 'Helvetica-Oblique';
+$up = -100;
+$ut = 50;
+$dw = 513;
+$diff = '';
+$enc = '';
+$desc = ['Flags' => 96,'FontBBox' => '[-170 -225 1116 931]','ItalicAngle' => -12,'Ascent' => 931,'Descent' => -225,'Leading' => 0,'CapHeight' => 718,'XHeight' => 523,'StemV' => 88,'StemH' => 76,'AvgWidth' => 513,'MaxWidth' => 1015,'MissingWidth' => 513];
+$cw = [0 => 278,1 => 278,2 => 278,3 => 278,4 => 278,5 => 278,6 => 278,7 => 278,8 => 278,9 => 278,10 => 278,11 => 278,12 => 278,13 => 278,14 => 278,15 => 278,16 => 278,17 => 278,18 => 278,19 => 278,20 => 278,21 => 278,22 => 278,23 => 278,24 => 278,25 => 278,26 => 278,27 => 278,28 => 278,29 => 278,30 => 278,31 => 278,32 => 278,33 => 278,34 => 355,35 => 556,36 => 556,37 => 889,38 => 667,39 => 191,40 => 333,41 => 333,42 => 389,43 => 584,44 => 278,45 => 333,46 => 278,47 => 278,48 => 556,49 => 556,50 => 556,51 => 556,52 => 556,53 => 556,54 => 556,55 => 556,56 => 556,57 => 556,58 => 278,59 => 278,60 => 584,61 => 584,62 => 584,63 => 556,64 => 1015,65 => 667,66 => 667,67 => 722,68 => 722,69 => 667,70 => 611,71 => 778,72 => 722,73 => 278,74 => 500,75 => 667,76 => 556,77 => 833,78 => 722,79 => 778,80 => 667,81 => 778,82 => 722,83 => 667,84 => 611,85 => 722,86 => 667,87 => 944,88 => 667,89 => 667,90 => 611,91 => 278,92 => 278,93 => 278,94 => 469,95 => 556,96 => 333,97 => 556,98 => 556,99 => 500,100 => 556,101 => 556,102 => 278,103 => 556,104 => 556,105 => 222,106 => 222,107 => 500,108 => 222,109 => 833,110 => 556,111 => 556,112 => 556,113 => 556,114 => 333,115 => 500,116 => 278,117 => 556,118 => 500,119 => 722,120 => 500,121 => 500,122 => 500,123 => 334,124 => 260,125 => 334,126 => 584,127 => 350,128 => 556,129 => 350,130 => 222,131 => 556,132 => 333,133 => 1000,134 => 556,135 => 556,136 => 333,137 => 1000,138 => 667,139 => 333,140 => 1000,141 => 350,142 => 611,143 => 350,144 => 350,145 => 222,146 => 222,147 => 333,148 => 333,149 => 350,150 => 556,151 => 1000,152 => 333,153 => 1000,154 => 500,155 => 333,156 => 944,157 => 350,158 => 500,159 => 667,160 => 278,161 => 333,162 => 556,163 => 556,164 => 556,165 => 556,166 => 260,167 => 556,168 => 333,169 => 737,170 => 370,171 => 556,172 => 584,173 => 333,174 => 737,175 => 333,176 => 400,177 => 584,178 => 333,179 => 333,180 => 333,181 => 556,182 => 537,183 => 278,184 => 333,185 => 333,186 => 365,187 => 556,188 => 834,189 => 834,190 => 834,191 => 611,192 => 667,193 => 667,194 => 667,195 => 667,196 => 667,197 => 667,198 => 1000,199 => 722,200 => 667,201 => 667,202 => 667,203 => 667,204 => 278,205 => 278,206 => 278,207 => 278,208 => 722,209 => 722,210 => 778,211 => 778,212 => 778,213 => 778,214 => 778,215 => 584,216 => 778,217 => 722,218 => 722,219 => 722,220 => 722,221 => 667,222 => 667,223 => 611,224 => 556,225 => 556,226 => 556,227 => 556,228 => 556,229 => 556,230 => 889,231 => 500,232 => 556,233 => 556,234 => 556,235 => 556,236 => 278,237 => 278,238 => 278,239 => 278,240 => 556,241 => 556,242 => 556,243 => 556,244 => 556,245 => 556,246 => 556,247 => 584,248 => 611,249 => 556,250 => 556,251 => 556,252 => 556,253 => 500,254 => 556,255 => 500];
 // --- EOF ---
