@@ -1,12 +1,12 @@
 <?php
 // TCPDF FONT FILE DESCRIPTION
-$type='core';
-$name='Symbol';
-$up=-100;
-$ut=50;
-$dw=587;
-$diff='';
-$enc='';
-$desc=array('Flags'=>4,'FontBBox'=>'[-180 -293 1090 1010]','ItalicAngle'=>0,'Ascent'=>1010,'Descent'=>-293,'Leading'=>0,'CapHeight'=>1010,'StemV'=>85,'StemH'=>92,'AvgWidth'=>587,'MaxWidth'=>1042,'MissingWidth'=>587);
-$cw=array(0=>587,1=>587,2=>587,3=>587,4=>587,5=>587,6=>587,7=>587,8=>587,9=>587,10=>587,11=>587,12=>587,13=>587,14=>587,15=>587,16=>587,17=>587,18=>587,19=>587,20=>587,21=>587,22=>587,23=>587,24=>587,25=>587,26=>587,27=>587,28=>587,29=>587,30=>587,31=>587,32=>250,33=>333,34=>713,35=>500,36=>549,37=>833,38=>778,39=>439,40=>333,41=>333,42=>500,43=>549,44=>250,45=>549,46=>250,47=>278,48=>500,49=>500,50=>500,51=>500,52=>500,53=>500,54=>500,55=>500,56=>500,57=>500,58=>278,59=>278,60=>549,61=>549,62=>549,63=>444,64=>549,65=>722,66=>667,67=>722,68=>612,69=>611,70=>763,71=>603,72=>722,73=>333,74=>631,75=>722,76=>686,77=>889,78=>722,79=>722,80=>768,81=>741,82=>556,83=>592,84=>611,85=>690,86=>439,87=>768,88=>645,89=>795,90=>611,91=>333,92=>863,93=>333,94=>658,95=>500,96=>500,97=>631,98=>549,99=>549,100=>494,101=>439,102=>521,103=>411,104=>603,105=>329,106=>603,107=>549,108=>549,109=>576,110=>521,111=>549,112=>549,113=>521,114=>549,115=>603,116=>439,117=>576,118=>713,119=>686,120=>493,121=>686,122=>494,123=>480,124=>200,125=>480,126=>549,127=>587,128=>587,129=>587,130=>587,131=>587,132=>587,133=>587,134=>587,135=>587,136=>587,137=>587,138=>587,139=>587,140=>587,141=>587,142=>587,143=>587,144=>587,145=>587,146=>587,147=>587,148=>587,149=>587,150=>587,151=>587,152=>587,153=>587,154=>587,155=>587,156=>587,157=>587,158=>587,159=>587,160=>750,161=>620,162=>247,163=>549,164=>167,165=>713,166=>500,167=>753,168=>753,169=>753,170=>753,171=>1042,172=>987,173=>603,174=>987,175=>603,176=>400,177=>549,178=>411,179=>549,180=>549,181=>713,182=>494,183=>460,184=>549,185=>549,186=>549,187=>549,188=>1000,189=>603,190=>1000,191=>658,192=>823,193=>686,194=>795,195=>987,196=>768,197=>768,198=>823,199=>768,200=>768,201=>713,202=>713,203=>713,204=>713,205=>713,206=>713,207=>713,208=>768,209=>713,210=>790,211=>790,212=>890,213=>823,214=>549,215=>250,216=>713,217=>603,218=>603,219=>1042,220=>987,221=>603,222=>987,223=>603,224=>494,225=>329,226=>790,227=>790,228=>786,229=>713,230=>384,231=>384,232=>384,233=>384,234=>384,235=>384,236=>494,237=>494,238=>494,239=>494,240=>587,241=>329,242=>274,243=>686,244=>686,245=>686,246=>384,247=>384,248=>384,249=>384,250=>384,251=>384,252=>494,253=>494,254=>494,255=>587);
+$type = 'core';
+$name = 'Symbol';
+$up = -100;
+$ut = 50;
+$dw = 587;
+$diff = '';
+$enc = '';
+$desc = ['Flags' => 4,'FontBBox' => '[-180 -293 1090 1010]','ItalicAngle' => 0,'Ascent' => 1010,'Descent' => -293,'Leading' => 0,'CapHeight' => 1010,'StemV' => 85,'StemH' => 92,'AvgWidth' => 587,'MaxWidth' => 1042,'MissingWidth' => 587];
+$cw = [0 => 587,1 => 587,2 => 587,3 => 587,4 => 587,5 => 587,6 => 587,7 => 587,8 => 587,9 => 587,10 => 587,11 => 587,12 => 587,13 => 587,14 => 587,15 => 587,16 => 587,17 => 587,18 => 587,19 => 587,20 => 587,21 => 587,22 => 587,23 => 587,24 => 587,25 => 587,26 => 587,27 => 587,28 => 587,29 => 587,30 => 587,31 => 587,32 => 250,33 => 333,34 => 713,35 => 500,36 => 549,37 => 833,38 => 778,39 => 439,40 => 333,41 => 333,42 => 500,43 => 549,44 => 250,45 => 549,46 => 250,47 => 278,48 => 500,49 => 500,50 => 500,51 => 500,52 => 500,53 => 500,54 => 500,55 => 500,56 => 500,57 => 500,58 => 278,59 => 278,60 => 549,61 => 549,62 => 549,63 => 444,64 => 549,65 => 722,66 => 667,67 => 722,68 => 612,69 => 611,70 => 763,71 => 603,72 => 722,73 => 333,74 => 631,75 => 722,76 => 686,77 => 889,78 => 722,79 => 722,80 => 768,81 => 741,82 => 556,83 => 592,84 => 611,85 => 690,86 => 439,87 => 768,88 => 645,89 => 795,90 => 611,91 => 333,92 => 863,93 => 333,94 => 658,95 => 500,96 => 500,97 => 631,98 => 549,99 => 549,100 => 494,101 => 439,102 => 521,103 => 411,104 => 603,105 => 329,106 => 603,107 => 549,108 => 549,109 => 576,110 => 521,111 => 549,112 => 549,113 => 521,114 => 549,115 => 603,116 => 439,117 => 576,118 => 713,119 => 686,120 => 493,121 => 686,122 => 494,123 => 480,124 => 200,125 => 480,126 => 549,127 => 587,128 => 587,129 => 587,130 => 587,131 => 587,132 => 587,133 => 587,134 => 587,135 => 587,136 => 587,137 => 587,138 => 587,139 => 587,140 => 587,141 => 587,142 => 587,143 => 587,144 => 587,145 => 587,146 => 587,147 => 587,148 => 587,149 => 587,150 => 587,151 => 587,152 => 587,153 => 587,154 => 587,155 => 587,156 => 587,157 => 587,158 => 587,159 => 587,160 => 750,161 => 620,162 => 247,163 => 549,164 => 167,165 => 713,166 => 500,167 => 753,168 => 753,169 => 753,170 => 753,171 => 1042,172 => 987,173 => 603,174 => 987,175 => 603,176 => 400,177 => 549,178 => 411,179 => 549,180 => 549,181 => 713,182 => 494,183 => 460,184 => 549,185 => 549,186 => 549,187 => 549,188 => 1000,189 => 603,190 => 1000,191 => 658,192 => 823,193 => 686,194 => 795,195 => 987,196 => 768,197 => 768,198 => 823,199 => 768,200 => 768,201 => 713,202 => 713,203 => 713,204 => 713,205 => 713,206 => 713,207 => 713,208 => 768,209 => 713,210 => 790,211 => 790,212 => 890,213 => 823,214 => 549,215 => 250,216 => 713,217 => 603,218 => 603,219 => 1042,220 => 987,221 => 603,222 => 987,223 => 603,224 => 494,225 => 329,226 => 790,227 => 790,228 => 786,229 => 713,230 => 384,231 => 384,232 => 384,233 => 384,234 => 384,235 => 384,236 => 494,237 => 494,238 => 494,239 => 494,240 => 587,241 => 329,242 => 274,243 => 686,244 => 686,245 => 686,246 => 384,247 => 384,248 => 384,249 => 384,250 => 384,251 => 384,252 => 494,253 => 494,254 => 494,255 => 587];
 // --- EOF ---
