@@ -1,12 +1,12 @@
 <?php
 // TCPDF FONT FILE DESCRIPTION
-$type='core';
-$name='Times-BoldItalic';
-$up=-100;
-$ut=50;
-$dw=501;
-$diff='';
-$enc='';
-$desc=array('Flags'=>96,'FontBBox'=>'[-200 -218 996 921]','ItalicAngle'=>-15,'Ascent'=>921,'Descent'=>-218,'Leading'=>0,'CapHeight'=>669,'XHeight'=>462,'StemV'=>121,'StemH'=>42,'AvgWidth'=>501,'MaxWidth'=>1000,'MissingWidth'=>501);
-$cw=array(0=>250,1=>250,2=>250,3=>250,4=>250,5=>250,6=>250,7=>250,8=>250,9=>250,10=>250,11=>250,12=>250,13=>250,14=>250,15=>250,16=>250,17=>250,18=>250,19=>250,20=>250,21=>250,22=>250,23=>250,24=>250,25=>250,26=>250,27=>250,28=>250,29=>250,30=>250,31=>250,32=>250,33=>389,34=>555,35=>500,36=>500,37=>833,38=>778,39=>278,40=>333,41=>333,42=>500,43=>570,44=>250,45=>333,46=>250,47=>278,48=>500,49=>500,50=>500,51=>500,52=>500,53=>500,54=>500,55=>500,56=>500,57=>500,58=>333,59=>333,60=>570,61=>570,62=>570,63=>500,64=>832,65=>667,66=>667,67=>667,68=>722,69=>667,70=>667,71=>722,72=>778,73=>389,74=>500,75=>667,76=>611,77=>889,78=>722,79=>722,80=>611,81=>722,82=>667,83=>556,84=>611,85=>722,86=>667,87=>889,88=>667,89=>611,90=>611,91=>333,92=>278,93=>333,94=>570,95=>500,96=>333,97=>500,98=>500,99=>444,100=>500,101=>444,102=>333,103=>500,104=>556,105=>278,106=>278,107=>500,108=>278,109=>778,110=>556,111=>500,112=>500,113=>500,114=>389,115=>389,116=>278,117=>556,118=>444,119=>667,120=>500,121=>444,122=>389,123=>348,124=>220,125=>348,126=>570,127=>350,128=>500,129=>350,130=>333,131=>500,132=>500,133=>1000,134=>500,135=>500,136=>333,137=>1000,138=>556,139=>333,140=>944,141=>350,142=>611,143=>350,144=>350,145=>333,146=>333,147=>500,148=>500,149=>350,150=>500,151=>1000,152=>333,153=>1000,154=>389,155=>333,156=>722,157=>350,158=>389,159=>611,160=>250,161=>389,162=>500,163=>500,164=>500,165=>500,166=>220,167=>500,168=>333,169=>747,170=>266,171=>500,172=>606,173=>333,174=>747,175=>333,176=>400,177=>570,178=>300,179=>300,180=>333,181=>576,182=>500,183=>250,184=>333,185=>300,186=>300,187=>500,188=>750,189=>750,190=>750,191=>500,192=>667,193=>667,194=>667,195=>667,196=>667,197=>667,198=>944,199=>667,200=>667,201=>667,202=>667,203=>667,204=>389,205=>389,206=>389,207=>389,208=>722,209=>722,210=>722,211=>722,212=>722,213=>722,214=>722,215=>570,216=>722,217=>722,218=>722,219=>722,220=>722,221=>611,222=>611,223=>500,224=>500,225=>500,226=>500,227=>500,228=>500,229=>500,230=>722,231=>444,232=>444,233=>444,234=>444,235=>444,236=>278,237=>278,238=>278,239=>278,240=>500,241=>556,242=>500,243=>500,244=>500,245=>500,246=>500,247=>570,248=>500,249=>556,250=>556,251=>556,252=>556,253=>444,254=>500,255=>444);
+$type = 'core';
+$name = 'Times-BoldItalic';
+$up = -100;
+$ut = 50;
+$dw = 501;
+$diff = '';
+$enc = '';
+$desc = ['Flags' => 96,'FontBBox' => '[-200 -218 996 921]','ItalicAngle' => -15,'Ascent' => 921,'Descent' => -218,'Leading' => 0,'CapHeight' => 669,'XHeight' => 462,'StemV' => 121,'StemH' => 42,'AvgWidth' => 501,'MaxWidth' => 1000,'MissingWidth' => 501];
+$cw = [0 => 250,1 => 250,2 => 250,3 => 250,4 => 250,5 => 250,6 => 250,7 => 250,8 => 250,9 => 250,10 => 250,11 => 250,12 => 250,13 => 250,14 => 250,15 => 250,16 => 250,17 => 250,18 => 250,19 => 250,20 => 250,21 => 250,22 => 250,23 => 250,24 => 250,25 => 250,26 => 250,27 => 250,28 => 250,29 => 250,30 => 250,31 => 250,32 => 250,33 => 389,34 => 555,35 => 500,36 => 500,37 => 833,38 => 778,39 => 278,40 => 333,41 => 333,42 => 500,43 => 570,44 => 250,45 => 333,46 => 250,47 => 278,48 => 500,49 => 500,50 => 500,51 => 500,52 => 500,53 => 500,54 => 500,55 => 500,56 => 500,57 => 500,58 => 333,59 => 333,60 => 570,61 => 570,62 => 570,63 => 500,64 => 832,65 => 667,66 => 667,67 => 667,68 => 722,69 => 667,70 => 667,71 => 722,72 => 778,73 => 389,74 => 500,75 => 667,76 => 611,77 => 889,78 => 722,79 => 722,80 => 611,81 => 722,82 => 667,83 => 556,84 => 611,85 => 722,86 => 667,87 => 889,88 => 667,89 => 611,90 => 611,91 => 333,92 => 278,93 => 333,94 => 570,95 => 500,96 => 333,97 => 500,98 => 500,99 => 444,100 => 500,101 => 444,102 => 333,103 => 500,104 => 556,105 => 278,106 => 278,107 => 500,108 => 278,109 => 778,110 => 556,111 => 500,112 => 500,113 => 500,114 => 389,115 => 389,116 => 278,117 => 556,118 => 444,119 => 667,120 => 500,121 => 444,122 => 389,123 => 348,124 => 220,125 => 348,126 => 570,127 => 350,128 => 500,129 => 350,130 => 333,131 => 500,132 => 500,133 => 1000,134 => 500,135 => 500,136 => 333,137 => 1000,138 => 556,139 => 333,140 => 944,141 => 350,142 => 611,143 => 350,144 => 350,145 => 333,146 => 333,147 => 500,148 => 500,149 => 350,150 => 500,151 => 1000,152 => 333,153 => 1000,154 => 389,155 => 333,156 => 722,157 => 350,158 => 389,159 => 611,160 => 250,161 => 389,162 => 500,163 => 500,164 => 500,165 => 500,166 => 220,167 => 500,168 => 333,169 => 747,170 => 266,171 => 500,172 => 606,173 => 333,174 => 747,175 => 333,176 => 400,177 => 570,178 => 300,179 => 300,180 => 333,181 => 576,182 => 500,183 => 250,184 => 333,185 => 300,186 => 300,187 => 500,188 => 750,189 => 750,190 => 750,191 => 500,192 => 667,193 => 667,194 => 667,195 => 667,196 => 667,197 => 667,198 => 944,199 => 667,200 => 667,201 => 667,202 => 667,203 => 667,204 => 389,205 => 389,206 => 389,207 => 389,208 => 722,209 => 722,210 => 722,211 => 722,212 => 722,213 => 722,214 => 722,215 => 570,216 => 722,217 => 722,218 => 722,219 => 722,220 => 722,221 => 611,222 => 611,223 => 500,224 => 500,225 => 500,226 => 500,227 => 500,228 => 500,229 => 500,230 => 722,231 => 444,232 => 444,233 => 444,234 => 444,235 => 444,236 => 278,237 => 278,238 => 278,239 => 278,240 => 500,241 => 556,242 => 500,243 => 500,244 => 500,245 => 500,246 => 500,247 => 570,248 => 500,249 => 556,250 => 556,251 => 556,252 => 556,253 => 444,254 => 500,255 => 444];
 // --- EOF ---
