@@ -1,12 +1,12 @@
 <?php
 // TCPDF FONT FILE DESCRIPTION
-$type='core';
-$name='Times-Italic';
-$up=-100;
-$ut=50;
-$dw=491;
-$diff='';
-$enc='';
-$desc=array('Flags'=>96,'FontBBox'=>'[-169 -217 1010 883]','ItalicAngle'=>-15.5,'Ascent'=>883,'Descent'=>-217,'Leading'=>0,'CapHeight'=>653,'XHeight'=>441,'StemV'=>76,'StemH'=>32,'AvgWidth'=>491,'MaxWidth'=>1000,'MissingWidth'=>491);
-$cw=array(0=>250,1=>250,2=>250,3=>250,4=>250,5=>250,6=>250,7=>250,8=>250,9=>250,10=>250,11=>250,12=>250,13=>250,14=>250,15=>250,16=>250,17=>250,18=>250,19=>250,20=>250,21=>250,22=>250,23=>250,24=>250,25=>250,26=>250,27=>250,28=>250,29=>250,30=>250,31=>250,32=>250,33=>333,34=>420,35=>500,36=>500,37=>833,38=>778,39=>214,40=>333,41=>333,42=>500,43=>675,44=>250,45=>333,46=>250,47=>278,48=>500,49=>500,50=>500,51=>500,52=>500,53=>500,54=>500,55=>500,56=>500,57=>500,58=>333,59=>333,60=>675,61=>675,62=>675,63=>500,64=>920,65=>611,66=>611,67=>667,68=>722,69=>611,70=>611,71=>722,72=>722,73=>333,74=>444,75=>667,76=>556,77=>833,78=>667,79=>722,80=>611,81=>722,82=>611,83=>500,84=>556,85=>722,86=>611,87=>833,88=>611,89=>556,90=>556,91=>389,92=>278,93=>389,94=>422,95=>500,96=>333,97=>500,98=>500,99=>444,100=>500,101=>444,102=>278,103=>500,104=>500,105=>278,106=>278,107=>444,108=>278,109=>722,110=>500,111=>500,112=>500,113=>500,114=>389,115=>389,116=>278,117=>500,118=>444,119=>667,120=>444,121=>444,122=>389,123=>400,124=>275,125=>400,126=>541,127=>350,128=>500,129=>350,130=>333,131=>500,132=>556,133=>889,134=>500,135=>500,136=>333,137=>1000,138=>500,139=>333,140=>944,141=>350,142=>556,143=>350,144=>350,145=>333,146=>333,147=>556,148=>556,149=>350,150=>500,151=>889,152=>333,153=>980,154=>389,155=>333,156=>667,157=>350,158=>389,159=>556,160=>250,161=>389,162=>500,163=>500,164=>500,165=>500,166=>275,167=>500,168=>333,169=>760,170=>276,171=>500,172=>675,173=>333,174=>760,175=>333,176=>400,177=>675,178=>300,179=>300,180=>333,181=>500,182=>523,183=>250,184=>333,185=>300,186=>310,187=>500,188=>750,189=>750,190=>750,191=>500,192=>611,193=>611,194=>611,195=>611,196=>611,197=>611,198=>889,199=>667,200=>611,201=>611,202=>611,203=>611,204=>333,205=>333,206=>333,207=>333,208=>722,209=>667,210=>722,211=>722,212=>722,213=>722,214=>722,215=>675,216=>722,217=>722,218=>722,219=>722,220=>722,221=>556,222=>611,223=>500,224=>500,225=>500,226=>500,227=>500,228=>500,229=>500,230=>667,231=>444,232=>444,233=>444,234=>444,235=>444,236=>278,237=>278,238=>278,239=>278,240=>500,241=>500,242=>500,243=>500,244=>500,245=>500,246=>500,247=>675,248=>500,249=>500,250=>500,251=>500,252=>500,253=>444,254=>500,255=>444);
+$type = 'core';
+$name = 'Times-Italic';
+$up = -100;
+$ut = 50;
+$dw = 491;
+$diff = '';
+$enc = '';
+$desc = ['Flags' => 96,'FontBBox' => '[-169 -217 1010 883]','ItalicAngle' => -15.5,'Ascent' => 883,'Descent' => -217,'Leading' => 0,'CapHeight' => 653,'XHeight' => 441,'StemV' => 76,'StemH' => 32,'AvgWidth' => 491,'MaxWidth' => 1000,'MissingWidth' => 491];
+$cw = [0 => 250,1 => 250,2 => 250,3 => 250,4 => 250,5 => 250,6 => 250,7 => 250,8 => 250,9 => 250,10 => 250,11 => 250,12 => 250,13 => 250,14 => 250,15 => 250,16 => 250,17 => 250,18 => 250,19 => 250,20 => 250,21 => 250,22 => 250,23 => 250,24 => 250,25 => 250,26 => 250,27 => 250,28 => 250,29 => 250,30 => 250,31 => 250,32 => 250,33 => 333,34 => 420,35 => 500,36 => 500,37 => 833,38 => 778,39 => 214,40 => 333,41 => 333,42 => 500,43 => 675,44 => 250,45 => 333,46 => 250,47 => 278,48 => 500,49 => 500,50 => 500,51 => 500,52 => 500,53 => 500,54 => 500,55 => 500,56 => 500,57 => 500,58 => 333,59 => 333,60 => 675,61 => 675,62 => 675,63 => 500,64 => 920,65 => 611,66 => 611,67 => 667,68 => 722,69 => 611,70 => 611,71 => 722,72 => 722,73 => 333,74 => 444,75 => 667,76 => 556,77 => 833,78 => 667,79 => 722,80 => 611,81 => 722,82 => 611,83 => 500,84 => 556,85 => 722,86 => 611,87 => 833,88 => 611,89 => 556,90 => 556,91 => 389,92 => 278,93 => 389,94 => 422,95 => 500,96 => 333,97 => 500,98 => 500,99 => 444,100 => 500,101 => 444,102 => 278,103 => 500,104 => 500,105 => 278,106 => 278,107 => 444,108 => 278,109 => 722,110 => 500,111 => 500,112 => 500,113 => 500,114 => 389,115 => 389,116 => 278,117 => 500,118 => 444,119 => 667,120 => 444,121 => 444,122 => 389,123 => 400,124 => 275,125 => 400,126 => 541,127 => 350,128 => 500,129 => 350,130 => 333,131 => 500,132 => 556,133 => 889,134 => 500,135 => 500,136 => 333,137 => 1000,138 => 500,139 => 333,140 => 944,141 => 350,142 => 556,143 => 350,144 => 350,145 => 333,146 => 333,147 => 556,148 => 556,149 => 350,150 => 500,151 => 889,152 => 333,153 => 980,154 => 389,155 => 333,156 => 667,157 => 350,158 => 389,159 => 556,160 => 250,161 => 389,162 => 500,163 => 500,164 => 500,165 => 500,166 => 275,167 => 500,168 => 333,169 => 760,170 => 276,171 => 500,172 => 675,173 => 333,174 => 760,175 => 333,176 => 400,177 => 675,178 => 300,179 => 300,180 => 333,181 => 500,182 => 523,183 => 250,184 => 333,185 => 300,186 => 310,187 => 500,188 => 750,189 => 750,190 => 750,191 => 500,192 => 611,193 => 611,194 => 611,195 => 611,196 => 611,197 => 611,198 => 889,199 => 667,200 => 611,201 => 611,202 => 611,203 => 611,204 => 333,205 => 333,206 => 333,207 => 333,208 => 722,209 => 667,210 => 722,211 => 722,212 => 722,213 => 722,214 => 722,215 => 675,216 => 722,217 => 722,218 => 722,219 => 722,220 => 722,221 => 556,222 => 611,223 => 500,224 => 500,225 => 500,226 => 500,227 => 500,228 => 500,229 => 500,230 => 667,231 => 444,232 => 444,233 => 444,234 => 444,235 => 444,236 => 278,237 => 278,238 => 278,239 => 278,240 => 500,241 => 500,242 => 500,243 => 500,244 => 500,245 => 500,246 => 500,247 => 675,248 => 500,249 => 500,250 => 500,251 => 500,252 => 500,253 => 444,254 => 500,255 => 444];
 // --- EOF ---
