@@ -1,12 +1,12 @@
 <?php
 // TCPDF FONT FILE DESCRIPTION
-$type='core';
-$name='ZapfDingbats';
-$up=-100;
-$ut=50;
-$dw=746;
-$diff='';
-$enc='';
-$desc=array('Flags'=>4,'FontBBox'=>'[-1 -143 981 820]','ItalicAngle'=>0,'Ascent'=>820,'Descent'=>-143,'Leading'=>0,'CapHeight'=>820,'StemV'=>90,'StemH'=>28,'AvgWidth'=>746,'MaxWidth'=>1016,'MissingWidth'=>746);
-$cw=array(0=>746,1=>746,2=>746,3=>746,4=>746,5=>746,6=>746,7=>746,8=>746,9=>746,10=>746,11=>746,12=>746,13=>746,14=>746,15=>746,16=>746,17=>746,18=>746,19=>746,20=>746,21=>746,22=>746,23=>746,24=>746,25=>746,26=>746,27=>746,28=>746,29=>746,30=>746,31=>746,32=>278,33=>974,34=>961,35=>974,36=>980,37=>719,38=>789,39=>790,40=>791,41=>690,42=>960,43=>939,44=>549,45=>855,46=>911,47=>933,48=>911,49=>945,50=>974,51=>755,52=>846,53=>762,54=>761,55=>571,56=>677,57=>763,58=>760,59=>759,60=>754,61=>494,62=>552,63=>537,64=>577,65=>692,66=>786,67=>788,68=>788,69=>790,70=>793,71=>794,72=>816,73=>823,74=>789,75=>841,76=>823,77=>833,78=>816,79=>831,80=>923,81=>744,82=>723,83=>749,84=>790,85=>792,86=>695,87=>776,88=>768,89=>792,90=>759,91=>707,92=>708,93=>682,94=>701,95=>826,96=>815,97=>789,98=>789,99=>707,100=>687,101=>696,102=>689,103=>786,104=>787,105=>713,106=>791,107=>785,108=>791,109=>873,110=>761,111=>762,112=>762,113=>759,114=>759,115=>892,116=>892,117=>788,118=>784,119=>438,120=>138,121=>277,122=>415,123=>392,124=>392,125=>668,126=>668,127=>746,128=>390,129=>390,130=>317,131=>317,132=>276,133=>276,134=>509,135=>509,136=>410,137=>410,138=>234,139=>234,140=>334,141=>334,142=>746,143=>746,144=>746,145=>746,146=>746,147=>746,148=>746,149=>746,150=>746,151=>746,152=>746,153=>746,154=>746,155=>746,156=>746,157=>746,158=>746,159=>746,160=>746,161=>732,162=>544,163=>544,164=>910,165=>667,166=>760,167=>760,168=>776,169=>595,170=>694,171=>626,172=>788,173=>788,174=>788,175=>788,176=>788,177=>788,178=>788,179=>788,180=>788,181=>788,182=>788,183=>788,184=>788,185=>788,186=>788,187=>788,188=>788,189=>788,190=>788,191=>788,192=>788,193=>788,194=>788,195=>788,196=>788,197=>788,198=>788,199=>788,200=>788,201=>788,202=>788,203=>788,204=>788,205=>788,206=>788,207=>788,208=>788,209=>788,210=>788,211=>788,212=>894,213=>838,214=>1016,215=>458,216=>748,217=>924,218=>748,219=>918,220=>927,221=>928,222=>928,223=>834,224=>873,225=>828,226=>924,227=>924,228=>917,229=>930,230=>931,231=>463,232=>883,233=>836,234=>836,235=>867,236=>867,237=>696,238=>696,239=>874,240=>746,241=>874,242=>760,243=>946,244=>771,245=>865,246=>771,247=>888,248=>967,249=>888,250=>831,251=>873,252=>927,253=>970,254=>918,255=>746);
+$type = 'core';
+$name = 'ZapfDingbats';
+$up = -100;
+$ut = 50;
+$dw = 746;
+$diff = '';
+$enc = '';
+$desc = ['Flags' => 4,'FontBBox' => '[-1 -143 981 820]','ItalicAngle' => 0,'Ascent' => 820,'Descent' => -143,'Leading' => 0,'CapHeight' => 820,'StemV' => 90,'StemH' => 28,'AvgWidth' => 746,'MaxWidth' => 1016,'MissingWidth' => 746];
+$cw = [0 => 746,1 => 746,2 => 746,3 => 746,4 => 746,5 => 746,6 => 746,7 => 746,8 => 746,9 => 746,10 => 746,11 => 746,12 => 746,13 => 746,14 => 746,15 => 746,16 => 746,17 => 746,18 => 746,19 => 746,20 => 746,21 => 746,22 => 746,23 => 746,24 => 746,25 => 746,26 => 746,27 => 746,28 => 746,29 => 746,30 => 746,31 => 746,32 => 278,33 => 974,34 => 961,35 => 974,36 => 980,37 => 719,38 => 789,39 => 790,40 => 791,41 => 690,42 => 960,43 => 939,44 => 549,45 => 855,46 => 911,47 => 933,48 => 911,49 => 945,50 => 974,51 => 755,52 => 846,53 => 762,54 => 761,55 => 571,56 => 677,57 => 763,58 => 760,59 => 759,60 => 754,61 => 494,62 => 552,63 => 537,64 => 577,65 => 692,66 => 786,67 => 788,68 => 788,69 => 790,70 => 793,71 => 794,72 => 816,73 => 823,74 => 789,75 => 841,76 => 823,77 => 833,78 => 816,79 => 831,80 => 923,81 => 744,82 => 723,83 => 749,84 => 790,85 => 792,86 => 695,87 => 776,88 => 768,89 => 792,90 => 759,91 => 707,92 => 708,93 => 682,94 => 701,95 => 826,96 => 815,97 => 789,98 => 789,99 => 707,100 => 687,101 => 696,102 => 689,103 => 786,104 => 787,105 => 713,106 => 791,107 => 785,108 => 791,109 => 873,110 => 761,111 => 762,112 => 762,113 => 759,114 => 759,115 => 892,116 => 892,117 => 788,118 => 784,119 => 438,120 => 138,121 => 277,122 => 415,123 => 392,124 => 392,125 => 668,126 => 668,127 => 746,128 => 390,129 => 390,130 => 317,131 => 317,132 => 276,133 => 276,134 => 509,135 => 509,136 => 410,137 => 410,138 => 234,139 => 234,140 => 334,141 => 334,142 => 746,143 => 746,144 => 746,145 => 746,146 => 746,147 => 746,148 => 746,149 => 746,150 => 746,151 => 746,152 => 746,153 => 746,154 => 746,155 => 746,156 => 746,157 => 746,158 => 746,159 => 746,160 => 746,161 => 732,162 => 544,163 => 544,164 => 910,165 => 667,166 => 760,167 => 760,168 => 776,169 => 595,170 => 694,171 => 626,172 => 788,173 => 788,174 => 788,175 => 788,176 => 788,177 => 788,178 => 788,179 => 788,180 => 788,181 => 788,182 => 788,183 => 788,184 => 788,185 => 788,186 => 788,187 => 788,188 => 788,189 => 788,190 => 788,191 => 788,192 => 788,193 => 788,194 => 788,195 => 788,196 => 788,197 => 788,198 => 788,199 => 788,200 => 788,201 => 788,202 => 788,203 => 788,204 => 788,205 => 788,206 => 788,207 => 788,208 => 788,209 => 788,210 => 788,211 => 788,212 => 894,213 => 838,214 => 1016,215 => 458,216 => 748,217 => 924,218 => 748,219 => 918,220 => 927,221 => 928,222 => 928,223 => 834,224 => 873,225 => 828,226 => 924,227 => 924,228 => 917,229 => 930,230 => 931,231 => 463,232 => 883,233 => 836,234 => 836,235 => 867,236 => 867,237 => 696,238 => 696,239 => 874,240 => 746,241 => 874,242 => 760,243 => 946,244 => 771,245 => 865,246 => 771,247 => 888,248 => 967,249 => 888,250 => 831,251 => 873,252 => 927,253 => 970,254 => 918,255 => 746];
 // --- EOF ---
