@@ -42,23 +42,24 @@
  */
 
 if (php_sapi_name() != 'cli') {
-  echo 'You need to run this command from console.';
-  exit(1);
+    echo 'You need to run this command from console.';
+    exit(1);
 }
 
-$tcpdf_include_dirs = array(realpath(dirname(__FILE__).'/../IXCTCPDF.php'), '/usr/share/php/tcpdf/IXCTCPDF.php', '/usr/share/tcpdf/IXCTCPDF.php', '/usr/share/php-tcpdf/IXCTCPDF.php', '/var/www/tcpdf/IXCTCPDF.php', '/var/www/html/tcpdf/IXCTCPDF.php', '/usr/local/apache2/htdocs/tcpdf/IXCTCPDF.php');
+$tcpdf_include_dirs = [realpath(dirname(__FILE__).'/../IXCTCPDF.php'), '/usr/share/php/tcpdf/IXCTCPDF.php', '/usr/share/tcpdf/IXCTCPDF.php', '/usr/share/php-tcpdf/IXCTCPDF.php', '/var/www/tcpdf/IXCTCPDF.php', '/var/www/html/tcpdf/IXCTCPDF.php', '/usr/local/apache2/htdocs/tcpdf/IXCTCPDF.php'];
 foreach ($tcpdf_include_dirs as $tcpdf_include_path) {
-	if (@file_exists($tcpdf_include_path)) {
-		require_once($tcpdf_include_path);
-		break;
-	}
+    if (@file_exists($tcpdf_include_path)) {
+        require_once($tcpdf_include_path);
+        break;
+    }
 }
 
 /**
  * Display help guide for this command.
  */
-function showHelp() {
-	$help = <<<EOD
+function showHelp()
+{
+    $help = <<<EOD
 tcpdf_addfont - command line tool to convert fonts for the TCPDF library.
 
 Usage: tcpdf_addfont.php [ options ] -i fontfile[,fontfile]...
@@ -121,20 +122,20 @@ Options:
 	-h
 	--help      Display this help and exit.
 EOD;
-	echo $help."\n\n";
-	exit(0);
+    echo $help."\n\n";
+    exit(0);
 }
 
 // remove the name of the executing script
 array_shift($argv);
 
 // no options chosen
-if (!is_array($argv)) {
-  showHelp();
+if (! is_array($argv)) {
+    showHelp();
 }
 
 // initialize the array of options
-$options = array('type'=>'', 'enc'=>'', 'flags'=>32, 'outpath'=>K_PATH_FONTS, 'platid'=>3, 'encid'=>1, 'addcbbox'=>false, 'link'=>false);
+$options = ['type' => '', 'enc' => '', 'flags' => 32, 'outpath' => K_PATH_FONTS, 'platid' => 3, 'encid' => 1, 'addcbbox' => false, 'link' => false];
 
 // short input options
 $sopt = '';
@@ -150,7 +151,7 @@ $sopt .= 'i:';
 $sopt .= 'h';
 
 // long input options
-$lopt = array();
+$lopt = [];
 $lopt[] = 'type:';
 $lopt[] = 'enc:';
 $lopt[] = 'flags:';
@@ -167,75 +168,75 @@ $inopt = getopt($sopt, $lopt);
 
 // import options (with some sanitization)
 foreach ($inopt as $opt => $val) {
-	switch ($opt) {
-		case 't':
-		case 'type': {
-			if (in_array($val, array('TrueTypeUnicode', 'TrueType', 'Type1', 'CID0JP', 'CID0KR', 'CID0CS', 'CID0CT'))) {
-				$options['type'] = $val;
-			}
-			break;
-		}
-		case 'e':
-		case 'enc': {
-			$options['enc'] = $val;
-			break;
-		}
-		case 'f':
-		case 'flags': {
-			$options['flags'] = intval($val);
-			break;
-		}
-		case 'o':
-		case 'outpath': {
-			$options['outpath'] = realpath($val);
-			if (substr($options['outpath'], -1) != '/') {
-				$options['outpath'] .= '/';
-			}
-			break;
-		}
-		case 'p':
-		case 'platid': {
-			$options['platid'] = min(max(1, intval($val)), 3);
-			break;
-		}
-		case 'n':
-		case 'encid': {
-			$options['encid'] = min(max(0, intval($val)), 10);
-			break;
-		}
-		case 'b':
-		case 'addcbbox': {
-			$options['addcbbox'] = true;
-			break;
-		}
-		case 'l':
-		case 'link': {
-			$options['link'] = true;
-			break;
-		}
-		case 'i':
-		case 'fonts': {
-			$options['fonts'] = explode(',', $val);
-			break;
-		}
-		case 'h':
-		case 'help':
-		default: {
-			showHelp();
-			break;
-		}
-	} // end of switch
+    switch ($opt) {
+        case 't':
+        case 'type': {
+            if (in_array($val, ['TrueTypeUnicode', 'TrueType', 'Type1', 'CID0JP', 'CID0KR', 'CID0CS', 'CID0CT'])) {
+                $options['type'] = $val;
+            }
+            break;
+        }
+        case 'e':
+        case 'enc': {
+            $options['enc'] = $val;
+            break;
+        }
+        case 'f':
+        case 'flags': {
+            $options['flags'] = intval($val);
+            break;
+        }
+        case 'o':
+        case 'outpath': {
+            $options['outpath'] = realpath($val);
+            if (substr($options['outpath'], -1) != '/') {
+                $options['outpath'] .= '/';
+            }
+            break;
+        }
+        case 'p':
+        case 'platid': {
+            $options['platid'] = min(max(1, intval($val)), 3);
+            break;
+        }
+        case 'n':
+        case 'encid': {
+            $options['encid'] = min(max(0, intval($val)), 10);
+            break;
+        }
+        case 'b':
+        case 'addcbbox': {
+            $options['addcbbox'] = true;
+            break;
+        }
+        case 'l':
+        case 'link': {
+            $options['link'] = true;
+            break;
+        }
+        case 'i':
+        case 'fonts': {
+            $options['fonts'] = explode(',', $val);
+            break;
+        }
+        case 'h':
+        case 'help':
+        default: {
+            showHelp();
+            break;
+        }
+    } // end of switch
 } // end of while loop
 
 if (empty($options['fonts'])) {
-	echo "ERROR: missing input fonts (try --help for usage)\n\n";
-	exit(2);
+    echo "ERROR: missing input fonts (try --help for usage)\n\n";
+    exit(2);
 }
 
 // check the output path
-if (!is_dir($options['outpath']) OR !is_writable($options['outpath'])) {
-	echo "ERROR: Can't write to ".$options['outpath']."\n\n";
-	exit(3);
+if (! is_dir($options['outpath']) or ! is_writable($options['outpath'])) {
+    echo "ERROR: Can't write to ".$options['outpath']."\n\n";
+    exit(3);
 }
 
 echo "\n>>> Converting fonts for TCPDF:\n";
@@ -246,19 +247,19 @@ echo '*** Output dir set to '.$options['outpath']."\n";
 $errors = false;
 
 foreach ($options['fonts'] as $font) {
-	$fontfile = realpath($font);
-	$fontname = IXCTCPDF_FONTS::addTTFfont($fontfile, $options['type'], $options['enc'], $options['flags'], $options['outpath'], $options['platid'], $options['encid'], $options['addcbbox'], $options['link']);
-	if ($fontname === false) {
-		$errors = true;
-		echo "--- ERROR: can't add ".$font."\n";
-	} else {
-		echo "+++ OK   : ".$fontfile.' added as '.$fontname."\n";
-	}
+    $fontfile = realpath($font);
+    $fontname = IXCTCPDF_FONTS::addTTFfont($fontfile, $options['type'], $options['enc'], $options['flags'], $options['outpath'], $options['platid'], $options['encid'], $options['addcbbox'], $options['link']);
+    if ($fontname === false) {
+        $errors = true;
+        echo "--- ERROR: can't add ".$font."\n";
+    } else {
+        echo "+++ OK   : ".$fontfile.' added as '.$fontname."\n";
+    }
 }
 
 if ($errors) {
-	echo "--- Process completed with ERRORS!\n\n";
-	exit(4);
+    echo "--- Process completed with ERRORS!\n\n";
+    exit(4);
 }
 
 echo ">>> Process successfully completed!\n\n";
